Add files via upload
java code to generate a random number between 1 to 100 and prompt the user to enter his guess until the number of attempts are over. If user guesses the right number then the code gets ended. If the guess is wrong user is prompt to enter another guess until attempt is over.